\name{xcmsSet-utils}
\Rdversion{1.1}
\alias{visualizeXcmsSet}

\title{Various utility functions for xcmsSet objects}

\description{
  These functions provide additional functionality for \code{xcmsSet}
  objects from the \code{xcms} package.
}
\usage{

visualizeXcmsSet(...)

}
\arguments{

  (in alphabetic order)

  \item{...}{
    Presently not used.
  }
}
\section{Data visualization}{

  \describe{

    \item{visualizeXcmsSet}{
      This function starts the \code{shiny} app to visualize
      \code{xcmsSet} objects present in the current workspace.
    }

  }
}

\value{
  Refer to the method and function description above for detailed
  information on the returned result object.
}
\author{
  Johannes Rainer.
}
\seealso{
  \code{\link{MSdata}}, \code{\link{MSslice}}, \code{\link{MSsliceList}}
}
\keyword{classes}

